Create bankTransactions.java
E5.8 Program that simulates a bank transaction record for 2 account a checking and a savings account. Ask for initial balances of both accounts and reject negative balances. Next, ask for the transactions that have taken place in each account. Transaction options are Deposit, Withdrawal, and Transfer between accounts. Next ask for the account in which the transactions must be associated to and REJECT transactions that overdraw an account. At the end print the balances of both accounts.